Implement base test fallback logic
It's possible to override an existing `base` test with a `custom`
version, if available. Otherwise the `base` test will be used. All
`base` tests are loaded by default, unless they are overwritten in the
`custom` folder.

Also specific custom tests that do not exist in the `base` folder will
be used for testing.